Changed: Decoupled the QC report from the main pipeline
